fix(cuda): widen 32-bit byte-size products on the dequant/upload path
Audit of the CUDA backend (managed `src/DotLLM.Cuda/` + native `native/kernels/`)
for the 32-bit index/pointer-arithmetic overflow class fixed on the CPU side in
issue #429 (`weightsQ8 + row * rowBytes` computed in `int`).

Native kernels: no confirmed sites. Every weight-row base (`w_row = weight +
(size_t)row * blocks_per_row * BYTES`) is already `size_t`-widened across
quantized_gemv / quantized_gemv_mmq / moe_grouped_gemv / iq2 / i2_s_gemv /
pq2_0_gemv / dequant* / embedding, and the checked-in PTX confirms the
widening survived codegen (mul.wide.s32 at those sites). Residual un-widened
multiplies are all intra-row block offsets (`w_row + sb * 144`), bounded by
one row's byte length.

Managed, CONFIRMED and fixed:

* `CudaKernels.LaunchDequantToF32`, F32 passthrough — `totalElements *
  sizeof(float)` evaluated in `int`, wrapping above 2^29 elements (a 2 GiB F32
  tensor). Reachable for a large embedding / LM head inside a 12-16 GB budget.
* `CudaKernels.LaunchDequantToF16`, F16 passthrough — `totalElements * 2`
  wraps above 2^30 elements (a 2 GiB F16 tensor). Both fed `(nuint)`, so a
  wrapped negative sign-extends to a ~16 EiB byte count.
* `CudaWeights.UploadAndDequant` — `int totalElements = outputDim * inputDim`
  wraps above 2^31 elements. Every byte size derived from it already widened
  to `long`; the element count itself was the remaining 32-bit product. Now
  `long`, with `checked((int)…)` at the two int-typed kernel entry points so a
  >2^31-element tensor throws instead of silently truncating. (Only reachable
  on cards far larger than the 12 GB RTX 3060 / 16 GB T4 — recorded, low
  priority. Llama 3.1 405B's LM head at 2.101e9 elements sits 2% under the
  int ceiling; the *2 / *4 sites above wrap long before it.)

Consistency-only (activation offsets, self-limiting; widened because the
sibling line two lines away already casts):
CudaTransformerModel.cs x2, HybridVulkanCudaTransformerModel.cs x1 —
`(nint)((seqLen - 1) * hiddenSize * h)` -> `(long)(seqLen - 1) * …`.

NOT verified on hardware: this box is AMD (no NVIDIA), so no CUDA kernel or
end-to-end run was possible, and nvcc is absent so the native PTX was not
recompiled (no native source was changed). Managed side builds clean
(DotLLM.Cuda and the full solution, 0 errors).

+++ src/DotLLM.Cuda/CudaKernels.cs
@@ -4728,7 +4728,10 @@ public sealed unsafe class CudaKernels : IDisposable
         switch (srcDtype)
         {
             case QuantizationType.F32:
-                CudaDriverApi.cuMemcpyDtoD_v2(dst, src, (nuint)(totalElements * sizeof(float))).ThrowOnError();
+                // (long) widening is load-bearing: `totalElements * 4` overflows int at
+                // totalElements > 2^29 (a 2 GiB F32 tensor — reachable for a large
+                // embedding / LM head inside a 12–16 GB budget).
+                CudaDriverApi.cuMemcpyDtoD_v2(dst, src, (nuint)((long)totalElements * sizeof(float))).ThrowOnError();
                 return;
 
             case QuantizationType.F16:
@@ -4888,8 +4891,10 @@ public sealed unsafe class CudaKernels : IDisposable
         switch (srcDtype)
         {
             case QuantizationType.F16:
-                // Already FP16, just copy
-                CudaDriverApi.cuMemcpyDtoD_v2(dst, src, (nuint)(totalElements * 2)).ThrowOnError();
+                // Already FP16, just copy.
+                // (long) widening is load-bearing: `totalElements * 2` overflows int at
+                // totalElements > 2^30 (a 2 GiB F16 tensor).
+                CudaDriverApi.cuMemcpyDtoD_v2(dst, src, (nuint)((long)totalElements * 2)).ThrowOnError();
                 return;
 
             case QuantizationType.F32:
